+++ app/controllers/sheets_controller.rb
@@ -77,7 +77,10 @@ class SheetsController < ApplicationController
   def create
     @sheet = current_user.sheets.where(project_id: @project.id, subject_id: @subject.id).new(sheet_params)
     if SheetTransaction.save_sheet!(@sheet, sheet_params, variables_params, current_user, request.remote_ip, 'sheet_create')
-      redirect_to [@sheet.project, @sheet], notice: 'Sheet was successfully created.'
+      notices = []
+      notices << 'Sheet was successfully created.'
+      notices << { label: 'Create another?', url:  new_data_entry_project_subject_path(@project, @subject, @sheet.design), class: 'fa-plus-square' }
+      redirect_to [@sheet.project, @sheet], notice: @sheet.design.repeated? ? notices : notices.first
     else
       render :new
     end