+++ db/migrate/20170209175452_add_repeated_to_designs.rb
@@ -1,0 +1,5 @@
+class AddRepeatedToDesigns < ActiveRecord::Migration[5.0]
+  def change
+    add_column :designs, :repeated, :boolean, null: false, default: false
+  end
+end